Split libcall handlers by domain
